+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Narrow type on initial assignment (#547)
 - Annotations in function bodies are not analyzed as evaluated (#564)
 - Invalid `cast`s show an error (#573)
+- Argument names are validated for subtypes (#562)
 ### Enhancements
 - Show 'narrowed from' in `reveal_type` (#550)
 - `--color-output` is enabled by default (#531)